refactor: make harmonizer a reuseable module instead, update some stufss
