+++ README.md
@@ -77,7 +77,7 @@ Generates a harmonized 16-color terminal palette. Supports all terminals and is 
 <summary>Unix Systems (Linux/macOS)</summary>
 
 Make sure to copy the following files from the `scripts/` folder to `~/.config/matugen/scripts/`:
-- `apply-term-colors.py`
+- `harmonizer.py`
 - `generate-term-colors.py`
 - `base-palette.json`
 
@@ -87,9 +87,10 @@ And then in your matugen config:
 # ...
 
 [templates.term-colors]
-input_path = './templates/term-colors.json'
-output_path = '~/.cache/matugen/term-colors.json'
-post_hook = 'python ~/.config/matugen/scripts/apply-term-colors.py'
+input_path = "./templates/term-colors.json"
+output_path = "~/.cache/matugen/term-colors.json"
+post_hook = "python ~/.config/matugen/scripts/generate-term-colors.py --colors ~/.cache/matugen/term-colors.json --apply --print"
+
 ```
 
 </details>
@@ -98,7 +99,7 @@ post_hook = 'python ~/.config/matugen/scripts/apply-term-colors.py'
 <summary>Windows</summary>
 
 Make sure to copy all four scripts from the `scripts/` folder to your matugen config's `scripts/` directory:
-- `apply-term-colors.py`
+- `harmonizer.py`
 - `generate-term-colors.py`
 - `base-palette.json`
 - `windows_term_post.ps1`
@@ -109,10 +110,9 @@ And then in your matugen config:
 # ...
 
 [templates.term-colors]
-input_path = './templates/term-colors.json'
-output_path = '~/.cache/matugen/term-colors.json'
-# windows must use absolute paths
-post_hook = 'python C:\\Users\\user\\AppData\\Roaming\\InioX\\matugen\\config\\scripts\\apply-term-colors.py'
+input_path = "./templates/term-colors.json"
+output_path = "~/.cache/matugen/term-colors.json"
+post_hook = "python ~/.config/matugen/scripts/generate-term-colors.py --colors ~/.cache/matugen/term-colors.json --apply --print"
 ```
 
 </details>
